refactor: move duplicate logic to new class
